compliance/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compliance-backend"
version = "0.1.0"
description = "Compliance & Audit-Readiness Service — Sprinto-equivalent, standalone or integrated with URIP"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # NEW-3 (CritFix-D) — Python 3.14 needs SQLAlchemy >= 2.0.40 for
    # `Mapped[uuid.UUID | None]` to type-resolve. 2.0.0 is too loose.
    "sqlalchemy[asyncio]>=2.0.40,<2.1",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # CRIT-005 (Z2.1) — python-jose removed (CVE-2024-33663 / 33664;
    # library is unmaintained). Replaced by pyjwt[crypto] declared below.
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "greenlet>=3.0.0",
    # PyJWT[crypto] — auditor_auth + auditor_service rely on pyjwt for
    # invitation tokens. Pinned to >=2.9 to track CVE-2024-33663/33664
    # fixes (the same bump CRIT-005 applied to the URIP backend).
    "pyjwt[crypto]>=2.9,<3",
    # cryptography — required by HKDF / Fernet helpers in middleware.
    "cryptography>=44.0.0,<45",
    # slowapi — rate-limit middleware in compliance_backend/middleware/rate_limit.py.
    "slowapi>=0.1.9,<0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["compliance_backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hatch.envs.default]
features = ["dev"]
